Hub_Plot Update 10/10
Script to do data analysis. So far, I have some functions to plot the
station locations via latitude and longitude, as well as functions to
return an array of all the stations and their respective time points.
It was too big at first so I had to divide by ten.